feat: :sparkles: 增加跳转参数

+++ src/utils/versionHandlers/8.ts
@@ -1,6 +1,10 @@
 export default function handler(db: IDBDatabase) {
-  db.deleteObjectStore('work');
-  db.deleteObjectStore('history');
+  if (db.objectStoreNames.contains('work')) {
+    db.deleteObjectStore('work');
+  }
+  if (db.objectStoreNames.contains('history')) {
+    db.deleteObjectStore('work');
+  }
   db.createObjectStore('work');
   db.createObjectStore('history').createIndex('utime', 'utime');
 }
